correct vote part

+++ src/main/java/com/pch777/bargains/web/VoteRestController.java
@@ -1,16 +1,25 @@
 package com.pch777.bargains.web;
 
+import java.security.Principal;
+import java.util.Date;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.pch777.bargains.exception.ResourceNotFoundException;
 import com.pch777.bargains.model.Vote;
+import com.pch777.bargains.model.VoteDto;
+import com.pch777.bargains.security.UserSecurity;
+import com.pch777.bargains.service.BargainService;
 import com.pch777.bargains.service.VoteService;
 
 import lombok.AllArgsConstructor;
@@ -21,30 +30,59 @@ import lombok.AllArgsConstructor;
 public class VoteRestController {
 
 	private VoteService voteService;
-	
+	private BargainService bargainService;
+	private UserSecurity userSecurity;
+
 	@GetMapping
 	public List<Vote> getAllVotes() {
 		return voteService.getAllVotes();
 	}
-	
+
 	@GetMapping("/{id}")
 	public ResponseEntity<Vote> getVoteById(@PathVariable Long id) {
-		return voteService
-				.getById(id)
+		return voteService.getById(id)
 				.map(vote -> ResponseEntity.ok(vote))
 				.orElse(ResponseEntity.notFound().build());
 	}
 
-	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> deleteVoteById(@PathVariable Long id) {
-		try {
-			voteService.deleteVoteById(id);
-			return ResponseEntity.noContent().build();
-		} catch (ResourceNotFoundException ex) {
-			return ResponseEntity.notFound().build();
-		}
+	@PostMapping("/{bargainId}")
+	public ResponseEntity<Object> vote(@PathVariable("bargainId") Long bargainId, @RequestBody VoteDto voteDto) {
+
+		return bargainService.getById(bargainId).map(bargain -> {
+			if(voteService.vote(voteDto, bargainId)) {
+			   return ResponseEntity.accepted().build();
+			}
+			
+			HttpStatus status = HttpStatus.BAD_REQUEST;
+			Map<String, Object> body = responseError(status,
+					"User has already voted on this bargain or user is an owner", "/api/votes");
+
+			return new ResponseEntity<Object>(body, status);
+
+		}).orElse(ResponseEntity.notFound().build());
+
 	}
-	
-	
-	
+
+	@DeleteMapping("/{id}")
+	public ResponseEntity<Object> deleteVoteById(@PathVariable Long id, Principal principal) {
+		return voteService.getById(id).map(v -> {
+				if(userSecurity.isOwnerOrAdmin(v.getUser().getEmail(), principal.getName())) {
+					voteService.deleteVoteById(id);
+					return ResponseEntity.noContent().build();
+				} else {
+	                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+	            }
+			}).orElse(ResponseEntity.notFound().build());				
+	}
+
+	public Map<String, Object> responseError(HttpStatus status, String message, String path) {
+		Map<String, Object> body = new LinkedHashMap<>();
+		body.put("timestamp", new Date());
+		body.put("status", status.value());
+		body.put("error", status.name());
+		body.put("message", message);
+		body.put("path", path);
+		return body;
+	}
+
 }
